fix, refactor : fix batch application component scan error and refactor timeout logic
