+++ domain/notice/src/main/kotlin/com/knet/dormitory/config/WebClientConfig.kt
@@ -2,10 +2,18 @@ package com.knet.dormitory.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.web.reactive.function.client.ExchangeStrategies
 import org.springframework.web.reactive.function.client.WebClient
 
 @Configuration
 class WebClientConfig {
     @Bean
-    fun webClient():WebClient = WebClient.create()
+    fun webClient(): WebClient {
+        val exchangeStrategies = ExchangeStrategies.builder()
+            .codecs { configurer -> configurer.defaultCodecs().maxInMemorySize(-1) }
+            .build()
+        return WebClient.builder()
+            .exchangeStrategies(exchangeStrategies)
+            .build()
+    }
 }
