+++ web/batch-server/src/main/kotlin/com/knet/batch/DormitoryBatchApplication.kt
@@ -3,7 +3,7 @@ package com.knet.batch
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
-@SpringBootApplication(scanBasePackages = ["com.knet.dormitory"])
+@SpringBootApplication(scanBasePackages = ["com.knet"])
 class DormitoryBatchApplication
 
 fun main(args: Array<String>) {
