--- web/batch-server/src/main/kotlin/com/knet/batch/config/ObjectMapperConfig.kt
@@ -1,18 +1,0 @@
-package com.knet.batch.config
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-
-@Configuration
-class ObjectMapperConfig {
-
-    @Bean
-    fun objectMapper(): ObjectMapper {
-        val objectMapper = ObjectMapper()
-        objectMapper.registerModule(JavaTimeModule())
-        return objectMapper
-    }
-}
